use this one!
at the end there is a section called "running the 9 logreg variations with pq16"- look at this one as it contains the full pipelines of logreg and xgboost and plotting methods, to understand how it works. above it , you have extra code of how to generate the labels, and how to do logreg with classes that aren't balanced (using models dictionary instead of models_5050).